third_party/blink/web_tests/platform/mac/external/wpt/editing/other/editable-state-and-focus-in-shadow-dom-in-designMode.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] SelectAll in the open shadow DOM
  promise_test: Unhandled rejection with value: object "TypeError: Invalid web code."
[FAIL] SelectAll in the <div contenteditable> in the open shadow DOM
  promise_test: Unhandled rejection with value: object "TypeError: Invalid web code."
[FAIL] SelectAll in the closed shadow DOM
  promise_test: Unhandled rejection with value: object "TypeError: Invalid web code."
[FAIL] SelectAll in the <div contenteditable> in the closed shadow DOM
  promise_test: Unhandled rejection with value: object "TypeError: Invalid web code."
[FAIL] Typing "A" after Collapse selection into text in <div contenteditable> in the open shadow DOM
  assert_equals: The shadow DOM shouldn't be modified after Collapse selection into text in <div contenteditable> in the open shadow DOM expected "<style>:focus { outline: 3px red solid; }</style><div>text<div contenteditable=\\"\\">Aeditable</div><object tabindex=\\"0\\">object</object><p tabindex=\\"0\\">paragraph</p></div>" but got "<style>:focus { outline: 3px red solid; }</style><div>text<div contenteditable=\\"\\">editable</div><object tabindex=\\"0\\">object</object><p tabindex=\\"0\\">paragraph</p></div>"
[FAIL] Typing "A" after Collapse selection into text in <div contenteditable> in the closed shadow DOM
  assert_equals: The shadow DOM shouldn't be modified after Collapse selection into text in <div contenteditable> in the closed shadow DOM expected "<style>:focus { outline: 3px red solid; }</style><div>text<div contenteditable=\\"\\">Aeditable</div><object tabindex=\\"0\\">object</object><p tabindex=\\"0\\">paragraph</p></div>" but got "<style>:focus { outline: 3px red solid; }</style><div>text<div contenteditable=\\"\\">editable</div><object tabindex=\\"0\\">object</object><p tabindex=\\"0\\">paragraph</p></div>"
Harness: the test ran to completion.
